'ex_03_imprima_soma_de_dois_numeros.py'

+++ secao_01_estrutura_sequencial/ex_03_imprima_soma_de_dois_numeros.py
@@ -14,6 +14,6 @@ Faça um Programa que peça dois números inteiros e imprima a soma.
 
 
 def imprima_a_soma_de_dois_numeros():
-    n1 = float(input('Digite um número: '))
-    n2 = float(input('Digite outro número: '))
+    n1 = int(input('Digite um número: '))
+    n2 = int(input('Digite outro número: '))
     print(f'A soma dos dois números informados é {n1+n2}')
